tolong diterima kak (ini hanya mengganti settings.py dan field model agar bisa di-deploy)

+++ book/models.py
@@ -5,7 +5,7 @@ class Book(models.Model):
     title = models.TextField(null=True, blank=True)
     description = models.TextField(null=True, blank=True)
     author = models.TextField(null=True, blank=True)
-    isbn10 = models.IntegerField(null=True, blank=True)
+    isbn10 = models.BigIntegerField(null=True, blank=True)
     isbn13 = models.TextField(null=True, blank=True)
     publish_date = models.TextField(null=True, blank=True)
     edition = models.IntegerField(null=True, blank=True)
